# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bookshelf API",
        "description": "API documentation for Bookshelf application",
        "contact": {
            "email": "support@test.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/chapters/{chapterId}/pages": {
            "get": {
                "tags": [
                    "Page"
                ],
                "summary": "Get all pages of a chapter",
                "operationId": "598a5dcdeb67d7c2c88b71f27796e1a7",
                "parameters": [
                    {
                        "name": "chapterId",
                        "in": "path",
                        "description": "ID of the chapter",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Chapter not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Page"
                ],
                "summary": "Create a new page in a chapter",
                "operationId": "df41eb2a878f02a5e7b0957671f61466",
                "parameters": [
                    {
                        "name": "chapterId",
                        "in": "path",
                        "description": "ID of the chapter",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "page_number",
                                    "content"
                                ],
                                "properties": {
                                    "page_number": {
                                        "type": "integer"
                                    },
                                    "content": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Page created successfully"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/chapters/{chapterId}/pages/{pageId}": {
            "get": {
                "tags": [
                    "Page"
                ],
                "summary": "Get a specific page of a chapter",
                "operationId": "b058abf4f50cef1bab68bac6b08a7f00",
                "parameters": [
                    {
                        "name": "chapterId",
                        "in": "path",
                        "description": "ID of the chapter",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "pageId",
                        "in": "path",
                        "description": "ID of the page",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Page"
                ],
                "summary": "Update a specific page of a chapter",
                "operationId": "70232896dd1e8a7227f6db8393b70916",
                "parameters": [
                    {
                        "name": "chapterId",
                        "in": "path",
                        "description": "ID of the chapter",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "pageId",
                        "in": "path",
                        "description": "ID of the page",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "page_number",
                                    "content"
                                ],
                                "properties": {
                                    "page_number": {
                                        "type": "integer"
                                    },
                                    "content": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Page updated successfully"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Page"
                ],
                "summary": "Delete a specific page of a chapter",
                "operationId": "c244c05af3438e10a656a38712ca9e8b",
                "parameters": [
                    {
                        "name": "chapterId",
                        "in": "path",
                        "description": "ID of the chapter",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "pageId",
                        "in": "path",
                        "description": "ID of the page",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Page deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearerAuth",
                "bearerFormat": "JWT"
            }
        }
    }
}